Removed redirect in api. Reorginized code. Moved default values to getWeather action.
